organize HKAudio files
